+++ tests/test_pinocchio.py
@@ -71,7 +71,3 @@ for name, description in DESCRIPTIONS.items():
             f"test_{name}",
             TestPinocchio.get_test_for_description(name),
         )
-
-
-if __name__ == "__main__":
-    unittest.main()